requirements.txt
streamlit==0.89.0
python-docx==0.8.10